module7/template_storage/code/main.py
import os
import tempfile # To create a temporary resized file before uploading
from google.cloud import storage
from wand.image import Image  # To resize image
# Add any imports that you may need, but make sure to update requirements.txt

def resize_image_storage(data, context):
    # TODO: Add logic here
    return