module7/template_storage/code/requirements.txt
google-cloud-storage==1.33.0
Wand==0.6.3